Rock, Paper, or Scissors
This program will allow two users play rock, paper, and scissors. Each win will be recorded and the results will be shown at the end of the match. Also, a single player has the option to play against the computer which is pretty fun. In addition, the scores continue to add up if the users choose to play multiple times.