Going Green
Description of program:  This program will prompt the user to enter the energy bills from January to December for the year prior to going green. Next, allow the user to enter the energy bills from January to December of the past year after going green.  The program should calculate the energy difference from the two years and display the two years worth of data, along with the savings.